Allow Sails session adapter (mongo) to reconnect
This fixes the
`Error occurred in session middleware :: 'MongoError: Topology was destroyed`
errors that occur when a Sails session is attempted to be created
when the underlying Mongo connection has been interrupted (e.g. if Mongo
is offline or had previously been offline or restarted, and its
connection affected).

Previously, this type of crash required a restart of the ReDBox
application because the topology/connection couldn't automatically
recover from its `DESTROYED` state.

This PR allows automatic recovery to occur on subsequent incoming
requests (or on a given request, if MongoDB is available again before
the topology timeout elapses).  For note, if this timeout does elapse,
such as if or when MongoDB is offline, users will still see a
network-related traceback, but this only persists until MongoDB becomes
available again.

This PR upgrades `connect-mongo` to its latest release version (v4.6.0) in
order to update the underlying `mongodb` package enabling better
topology management and automatic reconnection to mongo. To enable the
latest `connect-mongo` to be used, this PR upgrades Sails to v1.5.0 in
order to gain support for this newer package.  Additionally, since
`connect-mongo` now sets `mongodb` as a peer dependency, this PR adds
`mongodb` as a dependency to `redbox-portal`.